Support running Speedometer 3.0 benchmarks on Chrome
This version supports running Speedometer 3.0 benchmarks located at
https://github.com/WebKit/Speedometer on Chrome.

There are two hacks in this change, which can possibly be done better:

1) I didn't find yet an easy way for running-ng terminate Chrome automatically
after the benchmark iterations complete. As a workaround, I relied on timeouts
to kill Chrome, and used benchmark.is_passed() in minheap.py and runbms.py to
figure out if the timeout should be regarded as a success or failure.
This leads benchmarks to take longer to complete, because one needs to configure
a long enough timeout that guarantees completing the N iterations.

2) I didn't know what is the best way to enforce deleting Chrome's cache between
iterations. At the moment the hack I implemented is by adding
`rm -rf /tmp/chrome-expr-cache;` in get_log_prologue, and configuring the
benchmark to use the same cache directory.